Refactor publisher into cmd frontend and generic, browser-friendly backend
